requirements.txt
python-telegram-bot==20.3
Flask==2.3.2
pytz==2023.3
google-cloud-firestore==2.11.1
gunicorn==20.1.0